initial version of model predictor with onnx and libs

+++ exchange/file.go
@@ -9,6 +9,6 @@ type File struct {
 }
 
 type Prediction struct {
-	Label         int
+	Label         float32
 	Probabilities []float32
 }
